+++ tmtccmd/com/serial_dle.py
@@ -40,7 +40,8 @@ class SerialDleComIF(SerialComBase, ComInterface):
         self.__reception_buffer = deque()
         self.__polling_shutdown: Optional[threading.Event] = threading.Event()
 
-    def get_id(self) -> str:
+    @property
+    def id(self) -> str:
         return self.ser_cfg.com_if_id
 
     def initialize(self, args: any = None) -> any: